+++ packages/kysely-driver/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @powersync/kysely-driver
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [8cc1337]
+  - @journeyapps/powersync-sdk-common@1.5.1
+
 ## 0.1.6
 
 ### Patch Changes